fix warning fix warning

+++ tests/cpp-tests/Classes/UITest/CocoStudioGUITest/UIEditBoxTest.cpp
@@ -67,7 +67,7 @@ bool UIEditBoxTest::init()
         _editName->setPlaceHolder("Name:");
         _editName->setPlaceholderFontColor(Color3B::WHITE);
         _editName->setMaxLength(8);
-        _editName->setFontSize(editBoxSize.height/2);
+        _editName->setFontSize((int)editBoxSize.height/2);
         _editName->setText("v👐👊💝");
         _editName->setReturnType(ui::EditBox::KeyboardReturnType::DONE);
         _editName->setDelegate(this);
@@ -93,7 +93,7 @@ bool UIEditBoxTest::init()
         _editPassword->setMaxLength(6);
         _editPassword->setInputFlag(ui::EditBox::InputFlag::PASSWORD);
         _editPassword->setInputMode(ui::EditBox::InputMode::SINGLE_LINE);
-        _editPassword->setFontSize(editBoxSize.height/2);
+        _editPassword->setFontSize((int)editBoxSize.height/2);
         _editPassword->setDelegate(this);
         _editPassword->setVisible(true);
         addChild(_editPassword);
@@ -190,7 +190,7 @@ bool UIEditBoxTestToggleVisibility::init()
         addChild(_TTFShowEditReturn);
         
         
-        auto editBoxSize = Size(visibleSize.width - 100, visibleSize.height * 0.1);
+        auto editBoxSize = Size(visibleSize.width - 100, visibleSize.height * 0.1f);
         
         // top
         std::string pNormalSprite = "extensions/green_edit.png";
@@ -200,7 +200,7 @@ bool UIEditBoxTestToggleVisibility::init()
         _editName->setPlaceHolder("Name:");
         _editName->setPlaceholderFontColor(Color3B::WHITE);
         _editName->setMaxLength(8);
-        _editName->setFontSize(editBoxSize.height/2);
+        _editName->setFontSize((int)editBoxSize.height/2);
         _editName->setText("v👐👊💝");
         _editName->setReturnType(ui::EditBox::KeyboardReturnType::DONE);
         _editName->setDelegate(this);
@@ -225,7 +225,7 @@ bool UIEditBoxTestToggleVisibility::init()
         _editPassword->setMaxLength(6);
         _editPassword->setInputFlag(ui::EditBox::InputFlag::PASSWORD);
         _editPassword->setInputMode(ui::EditBox::InputMode::SINGLE_LINE);
-        _editPassword->setFontSize(editBoxSize.height/2);
+        _editPassword->setFontSize((float)editBoxSize.height/2);
         _editPassword->setDelegate(this);
         _editPassword->setVisible(true);
         addChild(_editPassword);
@@ -307,27 +307,27 @@ bool UIEditBoxTestTextHorizontalAlignment::init() {
     const auto glview = Director::getInstance()->getOpenGLView();
     const auto visibleOrigin = glview->getVisibleOrigin();
     const auto visibleSize = glview->getVisibleSize();
-    const auto editBoxSize = Size(visibleSize.width - 100, visibleSize.height * 0.1);
+    const auto editBoxSize = Size(visibleSize.width - 100, visibleSize.height * 0.1f);
   
     const auto createEditBox = [this, editBoxSize, visibleOrigin, visibleSize](const std::string& text,
                                       const TextHAlignment alignment,
                                       const int position_y) {
         ui::EditBox* editbox = ui::EditBox::create(editBoxSize + Size(0,40), ui::Scale9Sprite::create("extensions/green_edit.png"));
-        editbox->setPosition(Vec2(visibleOrigin.x+visibleSize.width/2-50, position_y));
+        editbox->setPosition(Vec2(visibleOrigin.x+visibleSize.width/2-50, (float)position_y));
         editbox->setFontColor(Color3B::RED);
         editbox->setPlaceHolder(text.c_str());
         editbox->setPlaceholderFontColor(Color3B::WHITE);
-        editbox->setPlaceholderFontSize(editBoxSize.height/2);
-        editbox->setFontSize(editBoxSize.height/2);
+        editbox->setPlaceholderFontSize((int)editBoxSize.height/2);
+        editbox->setFontSize((int)editBoxSize.height/2);
         editbox->setText(text.c_str());
         editbox->setTextHorizontalAlignment(alignment);
         editbox->setReturnType(ui::EditBox::KeyboardReturnType::DONE);
         addChild(editbox);
     };
   
-    createEditBox("horizontal left text", TextHAlignment::LEFT, visibleOrigin.y+visibleSize.height*3/4);
-    createEditBox("horizontal center text", TextHAlignment::CENTER, visibleOrigin.y+visibleSize.height/2);
-    createEditBox("horizontal right text", TextHAlignment::RIGHT, visibleOrigin.y+visibleSize.height/4);
+    createEditBox("horizontal left text", TextHAlignment::LEFT, (int)(visibleOrigin.y+visibleSize.height*3/4));
+    createEditBox("horizontal center text", TextHAlignment::CENTER, (int)(visibleOrigin.y+visibleSize.height/2));
+    createEditBox("horizontal right text", TextHAlignment::RIGHT, (int)(visibleOrigin.y+visibleSize.height/4));
   
     return true;
 }
@@ -341,15 +341,15 @@ bool UIEditBoxTestPressedAndDisabled::init() {
     auto glview = Director::getInstance()->getOpenGLView();
     auto visibleOrigin = glview->getVisibleOrigin();
     auto visibleSize = glview->getVisibleSize();
-    const auto editBoxSize = Size(visibleSize.width - 100, visibleSize.height * 0.1);
+    const auto editBoxSize = Size(visibleSize.width - 100, visibleSize.height * 0.1f);
 
     ui::EditBox* editbox = ui::EditBox::create(editBoxSize + Size(0,40), "extensions/yellow_edit.png", "extensions/orange_edit.png");
     editbox->setPosition(Vec2(visibleOrigin.x+visibleSize.width/2, visibleOrigin.y+visibleSize.height*3/4));
     editbox->setFontColor(Color3B::GREEN);
     editbox->setPlaceHolder("Test Pressed Image");
     editbox->setPlaceholderFontColor(Color3B::WHITE);
-    editbox->setPlaceholderFontSize(editBoxSize.height/2);
-    editbox->setFontSize(editBoxSize.height/2);
+    editbox->setPlaceholderFontSize((int)editBoxSize.height/2);
+    editbox->setFontSize((int)editBoxSize.height/2);
     editbox->setReturnType(ui::EditBox::KeyboardReturnType::DONE);
     addChild(editbox);
 
@@ -358,8 +358,8 @@ bool UIEditBoxTestPressedAndDisabled::init() {
     editbox->setFontColor(Color3B::GREEN);
     editbox->setPlaceHolder("Test Disabled Image - green");
     editbox->setPlaceholderFontColor(Color3B::WHITE);
-    editbox->setPlaceholderFontSize(editBoxSize.height/2);
-    editbox->setFontSize(editBoxSize.height/2);
+    editbox->setPlaceholderFontSize((int)editBoxSize.height/2);
+    editbox->setFontSize((int)editBoxSize.height/2);
     editbox->setReturnType(ui::EditBox::KeyboardReturnType::DONE);
     addChild(editbox);
     editbox->setEnabled(false);
@@ -369,8 +369,8 @@ bool UIEditBoxTestPressedAndDisabled::init() {
     editbox->setFontColor(Color3B::GREEN);
     editbox->setPlaceHolder("Disabled Gray effect");
     editbox->setPlaceholderFontColor(Color3B::WHITE);
-    editbox->setPlaceholderFontSize(editBoxSize.height/2);
-    editbox->setFontSize(editBoxSize.height/2);
+    editbox->setPlaceholderFontSize((int)editBoxSize.height/2);
+    editbox->setFontSize((int)editBoxSize.height/2);
     editbox->setReturnType(ui::EditBox::KeyboardReturnType::DONE);
     addChild(editbox);
     editbox->setEnabled(false);
